fix clippy

+++ src/tests/rust_guests/simpleguest/src/main.rs
@@ -612,7 +612,7 @@ fn spin_for_ms(fc: &FunctionCall) -> Result<Vec<u8>> {
     }
 
     // Calculate the actual number of milliseconds spun for, based on the counter and iterations per ms
-    let ms_spun = (counter / iterations_per_ms as u64) as u64;
+    let ms_spun = (counter / iterations_per_ms as u64);
     Ok(get_flatbuffer_result(ms_spun))
 }
